Extracted required host info
