+++ conf/script/src/build_system/cmd/setup/impl.py
@@ -1,5 +1,4 @@
 import platform
-import sys
 from pathlib import Path
 from typing import Final, Optional
 
@@ -12,47 +11,9 @@ import build_system.compiler.host
 import build_system.compiler.reqs.reqs
 import build_system.compiler.version
 import utils.error.cls_def
+from build_system.cmd.setup.required_host_info import detect_arch, fetch_os_family, fetch_supported_compiler_instances_by_os
 
 BUILD_DIR_PERMISSIONS: Final[int] = 0o770
-
-
-def fetch_os_name() -> str:
-    return platform.system().lower()
-
-
-def fetch_os_family() -> build_system.compiler.host.os_family.OSFamily:
-    # noinspection PyArgumentList
-    return build_system.compiler.host.os_family.OSFamily(fetch_os_name())
-
-
-def fetch_filtered_compilers_reqs_by_os(os_family: build_system.compiler.host.os_family.OSFamily) -> list[build_system.compiler.reqs.reqs.CompilerReqs]:
-    all_compilers_reqs = build_system.compiler.reqs.reqs.CompilerReqs.create_all_from_file()
-    return build_system.compiler.reqs.reqs.CompilerReqs.filter_by_os(all_compilers_reqs, os_family)
-
-
-def fetch_supported_compiler_instances_by_os(os_family: build_system.compiler.host.os_family.OSFamily) -> list[build_system.compiler.compiler_instance.CompilerInstance]:
-    filtered_compiler_reqs = fetch_filtered_compilers_reqs_by_os(os_family)
-    supported_compiler_instances: list[build_system.compiler.compiler_instance.CompilerInstance] = list()
-
-    for compiler_reqs in filtered_compiler_reqs:
-
-        if os_family in compiler_reqs.compiler_instance.os_families:
-
-            installed_compiler_instance = build_system.compiler.compiler_instance.CompilerInstance.create_from_installed_compiler(
-                compiler_family=compiler_reqs.compiler_instance.compiler_family, os_family=os_family)
-
-            if installed_compiler_instance.version >= compiler_reqs.compiler_instance.version:
-                supported_compiler_instances += installed_compiler_instance
-
-    return supported_compiler_instances
-
-
-def detect_arch() -> build_system.compiler.host.architecture.Architecture:
-    exclusive_max_word = sys.maxsize + 1
-    word_size = exclusive_max_word.bit_length()
-
-    # noinspection PyArgumentList
-    return build_system.compiler.host.architecture.Architecture(word_size)
 
 
 def assemble_build_types() -> list[build_system.cmd.setup.build_type.BuildType]:
@@ -117,7 +78,7 @@ def find_or_create_build_dir(root_dir: Optional[Path] = None) -> Path:
     return build_dir
 
 
-def setup(root_dir: Optional[Path] = None):
+def setup_build_system(root_dir: Optional[Path] = None):
     build_dir = find_or_create_build_dir(root_dir)
 
     all_build_subdir_names = generate_all_build_subdir_names()
